+++ src/katja_vmstats.erl
@@ -32,6 +32,7 @@
 % API
 -export([
   start/0,
+  stop/0,
   collect/1,
   get_timer/1,
   start_timer/2,
@@ -47,6 +48,15 @@ start() ->
   ok = application:start(protobuffs),
   ok = application:start(katja),
   ok = application:start(katja_vmstats),
+  ok.
+
+% @doc Stops the Katja VM Stats application and all of its dependencies. This is really only meant for usage inside the console.
+-spec stop() -> ok.
+stop() ->
+  ok = application:stop(katja_vmstats),
+  ok = application:stop(katja),
+  ok = application:stop(protobuffs),
+  ok = application:stop(noesis),
   ok.
 
 % @doc Collects the specified metrics and sends them to Riemann. Delegates to {@link katja_vmstats_collector:collect/1}.